// src/lib/orval/generated/model/experienceCompletionType.ts
/**
 * Generated by orval v7.13.1 🍺
 * Do not edit manually.
 * Gradegy API
 * Interactive documentation for the Gradegy API
 * OpenAPI spec version: 1.0
 */

/**
 * How the learner confirms completion of the experience.
 */
export type ExperienceCompletionType =
  (typeof ExperienceCompletionType)[keyof typeof ExperienceCompletionType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ExperienceCompletionType = {
  student: 'student',
  photo: 'photo',
  admin: 'admin',
  link: 'link',
} as const;